refactor: simplify worktree runtime integration
